study/research/cybercrime/qakbot/qakbot_analysis/tools/requirements.txt
pefile
pycryptodome
beautifulsoup4
requests_cache
